tests/innerhtml-02.html

<!DOCTYPE html>
<title>innerHTML in HTML</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://html5.org/specs/dom-parsing.html#innerhtml">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var tests = [
  [
    function() {
      var el = document.createElement("div");
      el.appendChild(document.createElement("xmp"))
        .appendChild(document.createElement("span"))
        .appendChild(document.createTextNode("<"));
      return el;
    },
    "<xmp><span>&lt;<\/span><\/xmp>"
  ],
  [
    function() {
      var el = document.createElement("xmp");
      el.appendChild(document.createElement("span"))
        .appendChild(document.createTextNode("<"));
      return el;
    },
    "<span>&lt;<\/span>"
  ],
  [
    function() {
      var el = document.createElement("xmp");
      el.appendChild(document.createTextNode("<"));
      return el;
    },
    "<"
  ],
  [
    function() {
      var el = document.createElement("div");
      el.appendChild(document.createElement("br"));
      return el;
    },
    "<br>"
  ]
];
tests.forEach(function(t) {
  test(function() {
    var el = t[0](), expected = t[1];
    assert_equals(el.innerHTML, expected);
  }, "Expected innerHTML: " + format_value(t[1]) + ".");
});
</script>
